requirements.txt
pandas
openpyxl
python-docx
matplotlib
seaborn
requests